Moved HOC into a separate file
